chore: FileUtilsTest에 있는 http version 관련 테스트 HttpUtilsTest로 이동
